requirements.txt
hdbcli>=2.19.0
pyswisseph>=2.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
numpy>=1.24.0
